C# Added, main menu added (WIP) and ragdoll fixes
I will start using C# more in the future now, maybe I will convert some scripts to C# as well.

A barely functional main menu was added were we can start the game, in the future the other options will have functions as well.

Player ragdoll has been slightly tweaked, the velocity threshold was lowered a bit as well for the impact sounds, I tweaked the ragdoll so that if we are walking on a angled floor, we instead of completely ignoring the floor direction, we fall perpendicular to it, relative to the velocity.